refactor(telegram): move markdown/HTML text transforms to markdown.rs, verbatim (#471 phase 1)
Second mechanical cut: md_to_html, command_md_to_html,
markdown_to_telegram_html, strip_html_tags, escape_html, format_inline,
convert_links, find_closing_marker and split_message move to a
dedicated markdown.rs. All are pure text transforms with no Bot API or
state dependency.

Pure move: byte-identical bodies, visibility widened to pub(crate) on
the two private helpers, glob re-export from handler keeps every call
site and test import stable.
