+++ src/channels/telegram/mod.rs
@@ -8,6 +8,7 @@ pub(crate) mod cowork;
 pub(crate) mod flow;
 pub(crate) mod follow_up_question;
 pub(crate) mod handler;
+pub(crate) mod markdown;
 pub(crate) mod raw_updates;
 pub(crate) mod reaction_prompt;
 pub(crate) mod rich;
